あれこれ修正

+++ src/domain/value_objects/man_hours.py
@@ -5,14 +5,14 @@ from util.calculator import get_hours_diff
 class ManHours:
     value: float
 
-    def __init__(self, value: int):
+    def __init__(self, value: float):
         if value is None:
             # raise ValueError("ManHoursは必須です。")
             value = 0  # デフォルト値を0に設定
         if value < 0:
             raise ValueError(f"ManHours`{value}`は負の数であってはいけません。")
         
-        self.value = value
+        self.value = float(value)
 
     @classmethod
     def from_notion_date(cls, date: NotionDate):
